Refactor runtime package API and encapsulation
