+++ runtime/runtime.go
@@ -17,18 +17,51 @@
 package runtime
 
 import (
-	"ballerina-lang-go/runtime/api"
+	"ballerina-lang-go/bir"
+	"ballerina-lang-go/runtime/internal/exec"
 	"ballerina-lang-go/runtime/internal/modules"
 )
 
+// Runtime represents a Ballerina runtime instance that owns a module registry
+// and is used as the execution context for interpreting BIR packages.
+type Runtime struct {
+	registry *modules.Registry
+}
+
+// ModuleInitializer is a function that can install modules (e.g. stdlibs) into
+// a runtime instance during its construction.
+type ModuleInitializer func(*Runtime)
+
+var moduleInitializers []ModuleInitializer
+
 // NewRuntime constructs a new runtime with an empty registry and runs all
 // registered module initializers.
-func NewRuntime() *api.Runtime {
-	rt := &api.Runtime{
-		Registry: modules.NewRegistry(),
+func NewRuntime() *Runtime {
+	rt := &Runtime{
+		registry: modules.NewRegistry(),
 	}
-	for _, init := range api.ModuleInitializers {
+	for _, init := range moduleInitializers {
 		init(rt)
 	}
 	return rt
+}
+
+// Interpret interprets a BIR package using this runtime instance.
+// It wraps the underlying interpreter with panic recovery and
+// returns any panic as an error value.
+func (rt *Runtime) Interpret(pkg bir.BIRPackage) (err error) {
+	exec.Interpret(pkg, rt.registry)
+	return err
+}
+
+// RegisterModuleInitializer registers a module initializer that will be invoked
+// for every newly created runtime.
+func RegisterModuleInitializer(init ModuleInitializer) {
+	moduleInitializers = append(moduleInitializers, init)
+}
+
+// RegisterExternFunction registers a native (extern) function implementation in
+// the given runtime instance so it can be called from interpreted BIR code.
+func RegisterExternFunction(rt *Runtime, orgName string, moduleName string, funcName string, impl func(args []any) (any, error)) {
+	rt.registry.RegisterExternFunction(orgName, moduleName, funcName, impl)
 }
